+++ cli/lib/default-config.js
@@ -12,7 +12,7 @@ module.exports = {
     pjax: true,
     predictivePrefetching: true,
     followRedirects: true,
-    usePercentage: false,
     pageJumpOffset: null,
     minimumConnection: "4g",
+    precacheURL: "",
 };