+++ src/core/pjax.ts
@@ -1,4 +1,3 @@
-import { hookup, message } from "../web_modules/broadcaster";
 import { env, uid } from "./env";
 import { sendPageView, setupGoogleAnalytics } from "./gtags.js";
 import { gaId, followRedirects, doPrefetching, pageJumpOffset, djinnjsOutDir, useServiceWorker } from "./config";
@@ -11,12 +10,11 @@ interface NavigaitonRequest {
     body?: string;
     title?: string;
     url: string;
-    history: "push" | "replace";
-    requestUid: string;
-    target: HTMLElement | null;
-    targetSelector: string;
-    tickets: string[];
-    customPageJumpOffset: number;
+    history?: "push" | "replace";
+    requestUid?: string;
+    targetSelector?: string;
+    tickets?: string[];
+    customPageJumpOffset?: number;
 }
 
 class Pjax {
@@ -25,7 +23,6 @@ class Pjax {
     private navigationRequestQueue: Array<NavigaitonRequest>;
     private io: IntersectionObserver;
     private serviceWorker: ServiceWorker;
-    private inboxUid: string;
 
     constructor() {
         this.state = {
@@ -51,8 +48,29 @@ class Pjax {
             localStorage.setItem("contentCache", `${Date.now()}`);
         }
 
-        /** Hookup Pjax's inbox */
-        this.inboxUid = hookup("pjax", this.inbox.bind(this));
+        document.addEventListener("pjax:revision", this.checkPageRevision);
+        document.addEventListener("pjax:load", (e: CustomEvent) => {
+            this.navigate({
+                url: e.detail.url,
+                history: e.detail?.history ?? "push",
+                targetSelector: e.detail?.selector,
+                tickets: e.detail?.tickets ?? [],
+                customPageJumpOffset: e.detail?.customPageJumpOffset,
+            });
+        });
+        document.addEventListener("pjax:continue", (e: CustomEvent) => {
+            this.swapPjaxContent(e.detail.requestUid);
+        });
+        document.addEventListener("pjax:init", () => {
+            this.worker = new Worker(`${location.origin}/${djinnjsOutDir}/pjax-worker.mjs`);
+            this.worker.onmessage = this.handleWorkerMessage.bind(this);
+            if (useServiceWorker && typeof navigator.serviceWorker !== "undefined") {
+                this.serviceWorker = navigator.serviceWorker.controller;
+                navigator.serviceWorker.onmessage = this.serviceWorkerInbox.bind(this);
+            }
+            this.checkPageRevision();
+            this.collectLinks();
+        });
 
         /** Prepare Google Analytics */
         setupGoogleAnalytics(gaId);
@@ -63,70 +81,6 @@ class Pjax {
         window.history.replaceState({ url: window.location.href }, document.title, window.location.href);
     }
 
-    /**
-     * The public inbox for the Pjax class. All incoming messages sent through the `Broadcaster` will be received here.
-     * @param data - the `MessageData` passed into the inbox by the `Broadcaster` class
-     */
-    private inbox(data): void {
-        const { type } = data;
-        switch (type) {
-            case "revision-check":
-                this.checkPageRevision();
-                break;
-            case "hijack-links":
-                this.collectLinks();
-                break;
-            case "load":
-                this.navigate(data.url, data?.history, data?.selector, data?.navRequestId, data?.tickets, data?.customPageJumpOffset);
-                break;
-            case "finalize-pjax":
-                this.updateHistory(data.title, data.url, data.history);
-                if (new RegExp("#").test(data.url)) {
-                    this.scrollToHash(data.url, data.customPageJumpOffset);
-                }
-                this.collectLinks();
-                this.checkPageRevision();
-                sendPageView(window.location.pathname, gaId);
-                if (doPrefetching) {
-                    this.prefetchLinks();
-                }
-                message({
-                    recipient: "pjax",
-                    type: "completed",
-                });
-                break;
-            case "css-ready":
-                this.swapPjaxContent(data.requestUid);
-                break;
-            case "prefetch":
-                if (doPrefetching) {
-                    this.prefetchLinks();
-                }
-                break;
-            case "init":
-                this.worker = new Worker(`${location.origin}/${djinnjsOutDir}/pjax-worker.mjs`);
-                this.worker.onmessage = this.handleWorkerMessage.bind(this);
-                if (useServiceWorker && typeof navigator.serviceWorker !== "undefined") {
-                    this.serviceWorker = navigator.serviceWorker.controller;
-                    navigator.serviceWorker.onmessage = this.serviceWorkerInbox.bind(this);
-                }
-                this.checkPageRevision();
-                /** Tell Pjax to hijack all viable links */
-                message({
-                    recipient: "pjax",
-                    type: "hijack-links",
-                });
-                /** Tell Pjax to prefetch links */
-                message({
-                    recipient: "pjax",
-                    type: "prefetch",
-                });
-                break;
-            default:
-                return;
-        }
-    }
-
     private serviceWorkerInbox(e: MessageEvent) {
         const { type } = e.data;
         switch (type) {
@@ -134,10 +88,8 @@ class Pjax {
                 let promptCount = parseInt(sessionStorage.getItem("prompts"));
                 promptCount = promptCount + 1;
                 sessionStorage.setItem("prompts", `${promptCount}`);
-                message({
-                    recipient: "user-input",
-                    type: "stale-notification",
-                });
+                const event = new CustomEvent("pjax:stale");
+                document.dispatchEvent(event);
                 break;
             case "cachebust":
                 sessionStorage.setItem("maxPrompts", `${e.data.max}`);
@@ -165,6 +117,18 @@ class Pjax {
             default:
                 break;
         }
+    }
+
+    private finalize(url: string, title: string, history: "push" | "replace", customPageJumpOffset: number) {
+        this.updateHistory(title, url, history);
+        if (new RegExp("#").test(url)) {
+            this.scrollToHash(url, customPageJumpOffset);
+        }
+        this.collectLinks();
+        this.checkPageRevision();
+        sendPageView(window.location.pathname, gaId);
+        const event = new CustomEvent("pjax:complete");
+        document.dispatchEvent(event);
     }
 
     /**
@@ -231,57 +195,37 @@ class Pjax {
     /**
      * Creates and sends a navigation request to the Pjax web worker and queues navigation request.
      */
-    private navigate(
-        url: string,
-        history: "push" | "replace" = "push",
-        selector: string = null,
-        navRequestId: string = null,
-        tickets: Array<string> = [],
-        customPageJumpOffset: number = null
-    ): void {
-        env.startPageTransition();
-        const requestUid = navRequestId || uid();
-        this.state.activeRequestUid = requestUid;
-        const navigationRequest: NavigaitonRequest = {
-            url: url,
-            history: history,
-            requestUid: requestUid,
-            target: document.body.querySelector(`[navigation-request-id="${requestUid}"]`) || null,
-            targetSelector: selector,
-            tickets: tickets,
-            customPageJumpOffset: customPageJumpOffset,
+    private navigate(request: NavigaitonRequest): void {
+        const defaultRequest: NavigaitonRequest = {
+            url: null,
+            history: "push",
+            requestUid: uid(),
+            targetSelector: null,
+            tickets: [],
+            customPageJumpOffset: null,
         };
+        env.startPageTransition();
+        const navigationRequest: NavigaitonRequest = Object.assign(defaultRequest, request);
+        this.state.activeRequestUid = navigationRequest.requestUid;
         this.navigationRequestQueue.push(navigationRequest);
         this.worker.postMessage({
             type: "pjax",
-            requestId: requestUid,
-            url: url,
+            requestId: navigationRequest.requestUid,
+            url: navigationRequest.url,
             currentUrl: location.href,
             followRedirects: followRedirects,
         });
     }
 
-    /**
-     * Handles the windows `popstate` event.
-     * @param e - the `PopStateEvent` object
-     */
-    private hijackPopstate(e: PopStateEvent): void {
+    private windowPopstateEvent: EventListener = (e: PopStateEvent) => {
         /** Only hijack the event when the `history.state` object contains a URL */
         if (e.state?.url) {
-            /** Tells the Pjax class to load the URL stored in this windows history.
-             * In order to preserve the timeline navigation the history will use `replace` instead of `push`.
-             */
-            message({
-                recipient: "pjax",
-                type: "load",
-                data: {
-                    url: e.state.url,
-                    history: "replace",
-                },
+            this.navigate({
+                url: e.state.url,
+                history: "replace",
             });
         }
-    }
-    private windowPopstateEvent: EventListener = this.hijackPopstate.bind(this);
+    };
 
     /**
      * Handles history manipulation by replacing or pushing the new state into the windows history timeline.
@@ -318,16 +262,23 @@ class Pjax {
         const navigationUid = uid();
         target.setAttribute("navigation-request-id", navigationUid);
         const customPageJumpOffset = target.getAttribute("page-jump-offset");
-        /** Tell Pjax to load the clicked elements page */
-        message({
-            recipient: "pjax",
-            type: "load",
-            data: {
-                url: target.href,
-                selector: target.getAttribute("pjax-view-id"),
-                navRequestId: navigationUid,
-                customPageJumpOffset: customPageJumpOffset ? parseInt(customPageJumpOffset) : null,
-            },
+        const offset = customPageJumpOffset ? parseInt(customPageJumpOffset) : null;
+        this.navigate({
+            url: target.href,
+            requestUid: navigationUid,
+            targetSelector: target.getAttribute("pjax-view-id"),
+            customPageJumpOffset: offset,
+        });
+    };
+
+    /**
+     * Called when the `mouseenter` event fires on a Pjax tracked anchor element.
+     */
+    private handleLinkPrefetch: EventListener = (e: Event) => {
+        const target = e.currentTarget as HTMLAnchorElement;
+        this.worker.postMessage({
+            type: "prefetch",
+            url: target.href,
         });
     };
 
@@ -337,13 +288,19 @@ class Pjax {
      * - any link with a `no-pjax` attribute
      * - any link with a `no-pjax` class
      * - any link with a `target` attribute
+     * - any link with a `download` attribute
      */
     private collectLinks(): void {
-        const unregisteredLinks = Array.from(document.body.querySelectorAll("a[href]:not([pjax-tracked]):not([no-pjax]):not([target]):not(.no-pjax)"));
+        const unregisteredLinks = Array.from(document.body.querySelectorAll("a[href]:not([pjax-tracked]):not([no-pjax]):not([target]):not(.no-pjax):not([download])"));
         if (unregisteredLinks.length) {
             unregisteredLinks.map((link: HTMLAnchorElement) => {
                 link.setAttribute("pjax-tracked", "true");
                 link.addEventListener("click", this.handleLinkClick);
+
+                /** Require at least a 3g connection & respect the users data saver setting */
+                if (doPrefetching && env.connection !== "2g" && env.connection !== "slow-2g" && !env.dataSaver) {
+                    link.addEventListener("mouseenter", this.handleLinkPrefetch);
+                }
             });
         }
     }
@@ -358,22 +315,20 @@ class Pjax {
      * @param error - the error message of the failed request
      */
     private handlePjaxResponse(requestId: string, status: string, url: string, body?: string, error?: string) {
-        const request = this.getNavigaitonRequest(requestId);
+        const request: NavigaitonRequest = this.getNavigaitonRequest(requestId);
         if (requestId === this.state.activeRequestUid) {
             if (status === "external") {
                 window.location.href = url;
             } else if (status === "hash-change") {
                 location.hash = url.match(/\#.*/g)[0].replace("#", "");
             } else if (status === "ok") {
-                /** Tells the runtime class to parse the incoming HTML for any new CSS files */
-                message({
-                    recipient: "runtime",
-                    type: "parse",
-                    data: {
+                const event = new CustomEvent("djinn:parse", {
+                    detail: {
                         body: body,
                         requestUid: requestId,
                     },
                 });
+                document.dispatchEvent(event);
                 request.body = body;
             } else {
                 console.error(`Failed to fetch page: ${url}. Server responded with: ${error}`);
@@ -392,7 +347,7 @@ class Pjax {
      * @param requestUid - the navigation request unique id
      */
     private swapPjaxContent(requestUid: string) {
-        const request = this.getNavigaitonRequest(requestUid);
+        const request: NavigaitonRequest = this.getNavigaitonRequest(requestUid);
         if (request.requestUid === this.state.activeRequestUid) {
             let selectors: Array<string> = [];
             let currentViews: Array<HTMLElement> = [];
@@ -473,27 +428,15 @@ class Pjax {
                 env.stopLoading(request.tickets[i]);
             }
 
-            message({
-                recipient: "pjax",
-                type: "finalize-pjax",
-                data: {
-                    url: request.url,
-                    title: tempDocument.title,
-                    history: request.history,
-                    customPageJumpOffset: request.customPageJumpOffset,
-                },
-            });
-            message({
-                recipient: "runtime",
-                type: "mount-components",
-            });
-            message({
-                recipient: "runtime",
-                type: "mount-inline-scripts",
-                data: {
+            this.finalize(request.url, tempDocument.title, request.history, request.customPageJumpOffset);
+            const mountEvent = new CustomEvent("djinn:mount-components");
+            document.dispatchEvent(mountEvent);
+            const scriptEvent = new CustomEvent("djinn:mount-scripts", {
+                detail: {
                     selectors: selectors,
                 },
             });
+            document.dispatchEvent(scriptEvent);
 
             env.endPageTransition();
         }
@@ -534,45 +477,6 @@ class Pjax {
         this.worker.postMessage({
             type: "revision-check",
             url: window.location.href,
-        });
-    }
-
-    /** Collect primary navigation links and tell the Pjax web worker to prefetch the pages. */
-    private prefetchLinks(): void {
-        /** Require a service worker & at least a 3g connection & respect the users data saver setting */
-        if (env.connection === "2g" || env.connection === "slow-2g" || !("serviceWorker" in navigator) || env.dataSaver) {
-            return;
-        }
-        const urls: Array<string> = [];
-
-        /** Header links */
-        const headerLinks = Array.from(document.body.querySelectorAll("header a[href]:not([target]):not([pjax-prefetched]):not(prevent-pjax):not(no-transition)"));
-        headerLinks.map((link: HTMLAnchorElement) => {
-            link.setAttribute("pjax-prefetched", "true");
-            urls.push(link.href);
-        });
-
-        /** All other navigation links */
-        const navLinks = Array.from(document.body.querySelectorAll("nav a[href]:not([target]):not([pjax-prefetched]):not(prevent-pjax):not(no-transition)"));
-        navLinks.map((link: HTMLAnchorElement) => {
-            link.setAttribute("pjax-prefetched", "true");
-            urls.push(link.href);
-        });
-
-        this.worker.postMessage({
-            type: "prefetch",
-            urls: urls,
-        });
-
-        /** Require at least a 4g connection while respecting the users data  */
-        if (env.connection === "3g") {
-            return;
-        }
-
-        const allLinks = Array.from(document.body.querySelectorAll("a[href]:not([target]):not([pjax-prefetched]):not(prevent-pjax):not(no-transition)"));
-        allLinks.map((link: HTMLAnchorElement) => {
-            link.setAttribute("pjax-prefetched", "true");
-            this.io.observe(link);
         });
     }